+++ src/SystemFolder/ControlPanels/SoundManager/index.ts
@@ -1,2 +1,4 @@
 export * from "./ClassicySoundManager";
 export * from "./ClassicySoundManagerContext";
+export * from "./ClassicySoundManagerProvider";
+export * from "./ClassicySoundManagerUtils";
