+++ src/SystemFolder/SystemResources/Window/ClassicyWindow.scss
@@ -197,7 +197,7 @@
     position: relative;
     display: flex;
     flex-direction: row;
-    height: calc(var(--window-control-size) * 1.5);
+    height: calc(var(--window-control-size) * 1.25);
     padding-bottom: var(--window-padding-size);
     text-align: center;
     border: none;
